Fix buzzer duration conflict w LED state machine
In no water state, the LED blinks at approx 100ms. The no water buzzer blocks the LED interrupts, which leads to more than 2000ms, interpreted as LED_ON or LED_OFF. This change is a simple fix/workaround to avoid this special case. Generally the Buzzer should not interrupt the LED detector - however, so far this was no issue, so will not make a big deal about it.

It is not clear to me, why this problem came up with a new machine today, and never before.